🎉 `adapter-ntchat`现在可用

+++ GenshinUID/genshinuid_update/restart.py
@@ -8,7 +8,7 @@ from pathlib import Path
 
 from ..utils.db_operation.db_operation import config_check
 
-bot_start = Path().cwd() / 'run.py'
+bot_start = Path().cwd() / 'bot.py'
 restart_sh_path = Path().cwd() / 'gs_restart.sh'
 update_log_path = Path(__file__).parent / 'update_log.json'
 
@@ -23,17 +23,17 @@ async def get_restart_sh(extra: str) -> str:
 
 
 async def restart_genshinuid(send_type: str, send_id: str) -> None:
-    pid = os.getpid()
     extra = ''
     if await config_check('UsePoetry'):
         extra = 'poetry run '
     extra += sys.executable
     restart_sh = await get_restart_sh(extra)
-    with open(restart_sh_path, "w", encoding="utf8") as f:
-        f.write(restart_sh)
-    if platform.system() == 'Linux':
-        os.system(f'chmod +x {str(restart_sh_path)}')
-        os.system(f'chmod +x {str(bot_start)}')
+    if not restart_sh_path.exists():
+        with open(restart_sh_path, "w", encoding="utf8") as f:
+            f.write(restart_sh)
+        if platform.system() == 'Linux':
+            os.system(f'chmod +x {str(restart_sh_path)}')
+            os.system(f'chmod +x {str(bot_start)}')
     now_time = time.strftime('%Y-%m-%d %H:%M:%S', time.localtime(time.time()))
     update_log = {
         'type': 'restart',
@@ -45,11 +45,9 @@ async def restart_genshinuid(send_type: str, send_id: str) -> None:
     with open(str(update_log_path), 'w', encoding='utf-8') as f:
         json.dump(update_log, f)
     if platform.system() == 'Linux':
-        subprocess.Popen(
-            f'kill -9 {pid} & {extra} {bot_start}',
-            shell=True,
-        )
+        os.execl(str(restart_sh_path), ' ')
     else:
+        pid = os.getpid()
         subprocess.Popen(
             f'taskkill /F /PID {pid} & {extra} {bot_start}',
             shell=True,
